requirements.txt
requests
bs4
cloudscraper
tldextract
python-Levenshtein